requirements.txt
# # requirements.txt

# streamlit==1.24.1
# sounddevice==0.4.6
# scipy==1.10.1
# openai==0.27.8
# whisper==1.0
# pinecone-client==2.2.1
# python-dotenv==1.0.0
# transformers==4.27.4
# torch==2.0.1
# sentencepiece==0.1.99

# # If you need to parse PDFs or docs:
# pdfplumber==0.9.0
# python-docx==0.8.11


# requirements.txt
streamlit
sounddevice
scipy
openai
whisper
pinecone
python-dotenv
transformers
torch
sentencepiece

# If you need to parse PDFs or docs:
pdfplumber
python-docx
